requirements.txt
streamlit
numpy
pandas
tensorflow
opencv-python
Pillow